modules/testing/integration-tests/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
attrs==23.2.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.202
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.2
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v6==2.0.3
    # via aws-cdk-lib
aws-cdk-integ-tests-alpha==2.149.0a0
    # via -r requirements.in
aws-cdk-lib==2.149.0
    # via
    #   -r requirements.in
    #   aws-cdk-integ-tests-alpha
boto3==1.34.144
    # via -r requirements.in
botocore==1.34.144
    # via
    #   boto3
    #   s3transfer
cattrs==23.2.3
    # via jsii
constructs==10.3.0
    # via
    #   -r requirements.in
    #   aws-cdk-integ-tests-alpha
    #   aws-cdk-lib
exceptiongroup==1.2.2
    # via cattrs
importlib-resources==6.4.0
    # via jsii
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsii==1.101.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-integ-tests-alpha
    #   aws-cdk-lib
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-integ-tests-alpha
    #   aws-cdk-lib
    #   constructs
    #   jsii
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   jsii
s3transfer==0.10.2
    # via boto3
six==1.16.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-integ-tests-alpha
    #   aws-cdk-lib
    #   constructs
    #   jsii
typing-extensions==4.12.2
    # via
    #   cattrs
    #   jsii
urllib3==1.26.19
    # via botocore
zipp==3.19.2
    # via importlib-resources
